+++ config/routes.rb
@@ -2,6 +2,8 @@ Slice::Application.routes.draw do
 
   mount MailPreview => 'mail_view' if Rails.env.development?
 
+  resources :comments
+
   resources :exports do
     member do
       post :mark_unread