+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Selected checkbox and radio responses are now highlighted
 - Sheet coverage added to give an approximate of how much of the sheet has been filled out
 - Documents and Links can now have their categories renamed *en masse*
+- Project Editors, Viewers, and Site Members can now comment on sheets
 
 ### Bug Fix
 - `Include missing` is now always displayed on the design report